Adicionadas funcionalidades de fornecedores e, movimentaÃçoes de estoqueorreções de bugs

+++ systemPOO/src/dao/interfaces/StorageDAO.java
@@ -3,7 +3,7 @@ package dao.interfaces;
 import entities.Product;
 
 public interface StorageDAO {
-	    void addProduct(Product product, int quantity) throws Exception;
-	    void removeProduct(Product product, int quantity) throws Exception;
-	    int checkProductQuantity(Product product) throws Exception;
-	}
+	void addProduct(Product product, Double quantity) throws Exception;
+	void removeProduct(Product product, Double quantity) throws Exception;
+    int checkProductQuantity(Product product) throws Exception;
+}
